implemented jwt security

+++ src/main/java/com/kibo/survey/business/abstracts/RatingService.java
@@ -17,6 +17,6 @@ public interface RatingService {
 
     Result addRatingList(List<RequestRatingDto> requestRatingDtos);
 
-    DataResult<Double> getAverageRatingByQuestionId(int questionId);
+    DataResult<Float> getAverageRatingByQuestionId(int questionId);
 
 }
